Added main and made some changes to adc.h and adc.c

+++ Assgn9/adc.h
@@ -14,7 +14,7 @@
 
 void initADC(void);
 uint16_t readADCFlag(void);
-uint16_t readADC(void);
+uint8_t readADC(void);
 void ADC14_IRQHandler(void);
 
 #endif /* ADC_H_ */
